core/delayed: moved to core/swapper
