+++ core/swapper/delayed/delayed_suite_test.go
@@ -1,0 +1,13 @@
+package delayed_test
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+)
+
+func TestDelayed(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Delayed Suite")
+}
